+++ R/tar_glimpse.R
@@ -9,17 +9,9 @@
 #'   renders faster. Also, `tar_glimpse()` omits functions and other global
 #'   objects by default (but you can include them with `targets_only = FALSE`).
 #' @return A `visNetwork` HTML widget object.
-#' @inheritParams tar_validate
+#' @inheritParams tar_network
 #' @param targets_only Logical, whether to restrict the output to just targets
 #'   (`FALSE`) or to also include global functions and objects.
-#' @param allow Optional, define the set of allowable vertices in the graph.
-#'   Set to `NULL` to allow all vertices in the pipeline and environment
-#'   (default). Otherwise, you can supply symbols or
-#'   `tidyselect` helpers like [starts_with()].
-#' @param exclude Optional, define the set of exclude vertices from the graph.
-#'   Set to `NULL` to exclude no vertices.
-#'   Otherwise, you can supply symbols or `tidyselect`
-#'   helpers like [starts_with()].
 #' @param level_separation Numeric of length 1,
 #'   `levelSeparation` argument of `visNetwork::visHierarchicalLayout()`.
 #'   Controls the distance between hierarchical levels.
@@ -50,6 +42,8 @@
 #' }
 tar_glimpse <- function(
   targets_only = TRUE,
+  names = NULL,
+  shortcut = FALSE,
   allow = NULL,
   exclude = ".Random.seed",
   level_separation = NULL,
@@ -75,6 +69,8 @@ tar_glimpse <- function(
   targets_arguments <- list(
     path_store = store,
     targets_only = targets_only,
+    names_quosure = rlang::enquo(names),
+    shortcut = shortcut,
     allow_quosure = rlang::enquo(allow),
     exclude_quosure = rlang::enquo(exclude),
     level_separation = level_separation,
@@ -95,6 +91,8 @@ tar_glimpse_inner <- function(
   pipeline,
   path_store,
   targets_only,
+  names_quosure,
+  shortcut,
   allow_quosure,
   exclude_quosure,
   level_separation,
@@ -103,16 +101,19 @@ tar_glimpse_inner <- function(
 ) {
   meta <- meta_init(path_store = path_store)
   progress <- progress_init(path_store = path_store)
+  names <- tar_tidyselect_eval(names_quosure, pipeline_get_names(pipeline))
   network <- glimpse_init(
     pipeline = pipeline,
     meta = meta,
-    progress = progress
+    progress = progress,
+    targets_only = targets_only,
+    names = names,
+    shortcut = shortcut,
+    allow = allow_quosure,
+    exclude = exclude_quosure
   )
   visual <- visnetwork_init(
     network = network,
-    targets_only = targets_only,
-    allow = allow_quosure,
-    exclude = exclude_quosure,
     level_separation = level_separation,
     degree_from = degree_from,
     degree_to = degree_to